해킹과 휘두르기,투척 공격의 애니메이션을 추가
해킹과 휘두르기, 투척 공격의 상태를 변수로 추가하고 변수의 값을 이용하여 애니메이션을 해당하는 상태가 되면 출력되도록 하였다. 또한 공격의 경우 공격도중에 이동을 할 수 없도록 함. 투척 공격은 모션 도중에 나가도록 하고, 휘두르기 공격의 공격 이팩트를 이동 및 회전시켜 싱크를 맞췄다.

+++ AvoidHimClient/Sound.cpp
@@ -23,7 +23,7 @@ void Sound::Play(bool _loop) {
 
 void Sound::SetSpeed(float _speed) {
 	// 지정된 샘플레이트를 벗어나게 하지 말것
-	pSecondaryBuffer->SetFrequency(sampleRate * _speed);
+	pSecondaryBuffer->SetFrequency(DWORD(sampleRate * _speed));
 }
 
 void Sound::Stop() {
